Fix issue with 0 being a valid coordinate value in product metadata
A recent fix removed "0" from the coordinates as there appeared in
some product metadata as a Z coordinate value for the scene
footprint. I mistakenly assumed that, because of the precision of the
x and y values, that there wouldn't be a "0" valid coordinate besides
in the Z value position; apparently that is not the case, and I've
found very few where '0' does actually represent a x or y coordinate.